+++ R/analyze.R
@@ -3,8 +3,7 @@
 #' Analyze objects. See the documentation for your object's class:
 #' \itemize{
 #'  \item{\link[=analyze.stanreg]{analyze.stanreg}}
-#'  \item{\link[=analyze.merModLmerTest]{analyze.merModLmerTest}}
-#'  \item{\link[=analyze.merMod]{analyze.merMod}}
+#'  \item{\link[=analyze.lmerModLmerTest]{analyze.merModLmerTest}}
 #'  \item{\link[=analyze.glmerMod]{analyze.glmerMod}}
 #'  }
 #'